refactor: split images
